Extension for blind
Here is a group of 3 extensions for blind people that helps them access features with 6 shortcuts. They can navigate websites, recognize links, and get notified when entering or leaving a site. A shortcut opens a textbox that pronounces letters typed and searches for words. It reads selected sentences and describes search results with shortcuts. Shortcuts are customizable.